… vendor/github.com/lucy/runewidth/LICENSE
… vendor/github.com/lucy/runewidth/compare/compare.go
… vendor/github.com/lucy/runewidth/gen.go
… vendor/github.com/lucy/runewidth/override.go
… vendor/github.com/lucy/runewidth/runewidth.go
… vendor/github.com/lucy/runewidth/tables.go
… vendor/github.com/mattn/go-runewidth/LICENSE
… vendor/github.com/mattn/go-runewidth/runewidth.go
… vendor/github.com/mattn/go-runewidth/runewidth_js.go
… vendor/github.com/mattn/go-runewidth/runewidth_posix.go
… vendor/github.com/mattn/go-runewidth/runewidth_windows.go
… vendor/golang.org/x/text/internal/gen/LICENSE
… vendor/golang.org/x/text/internal/gen/code.go
… vendor/golang.org/x/text/internal/gen/gen.go
… vendor/golang.org/x/text/internal/triegen/LICENSE
… vendor/golang.org/x/text/internal/triegen/compact.go
… vendor/golang.org/x/text/internal/triegen/print.go
… vendor/golang.org/x/text/internal/triegen/triegen.go
… vendor/golang.org/x/text/internal/ucd/LICENSE
… vendor/golang.org/x/text/internal/ucd/ucd.go
… vendor/golang.org/x/text/unicode/cldr/LICENSE
… vendor/golang.org/x/text/unicode/cldr/base.go
… vendor/golang.org/x/text/unicode/cldr/cldr.go
… vendor/golang.org/x/text/unicode/cldr/collate.go
… vendor/golang.org/x/text/unicode/cldr/decode.go
… vendor/golang.org/x/text/unicode/cldr/makexml.go
… vendor/golang.org/x/text/unicode/cldr/resolve.go
… vendor/golang.org/x/text/unicode/cldr/slice.go
… vendor/golang.org/x/text/unicode/cldr/xml.go
… vendor/manifest