@y0ssar1an 2017-04-05
Thanks for benchmarking this! I was concerned this would be slow, but didn't see an alternative since `lucy/runewidth` has gone away.